Move locale code to its own module
