+++ rust/agama-dbus-server/src/lib.rs
@@ -1,6 +1,4 @@
 pub mod error;
-pub mod helpers;
-pub mod keyboard;
 pub mod locale;
 pub mod network;
 pub mod questions;
