+++ rust/agama-dbus-server/src/locale/helpers.rs
